fastapi_app/requirements.txt
anthropic==0.40.0
pandas==2.2.3
pydantic==2.7.4
pydantic-settings==2.4.0
pydantic_core==2.18.4
pymongo==4.10.1
python-multipart==0.0.20
requests==2.32.3


httpx==0.27.0
fastapi==0.111.0
uvicorn==0.29.0
prometheus-client==0.20.0
opentelemetry-distro==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-exporter-otlp==1.24.0
